pyproject.toml
[project]
name = "mcp-sql-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.26.0",
    "ollama>=0.6.1",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
]
